migrate useEffect back to Overview module
