+++ client/App.jsx
@@ -1,27 +1,16 @@
 import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
-import axios from 'axios';
 
 import Overview from './components/overview/Overview';
 import QandA from './components/Q&A/Q&A';
 import Reviews from './components/reviews/Reviews';
 
 const App = () => {
-  const [productsList, setProductsList] = useState([]);
   const [productId, setProductId] = useState(0);
   useEffect(() => {
     // TODO: Set product id based on the URL
     setProductId(20100);
   });
-
-  useEffect(() => {
-    async function fetchProducts() {
-      const results = await axios.get('/products');
-      setProductsList(results.data);
-      setProductId(results.data[0].id); // sets first item as default product
-    }
-    return fetchProducts();
-  }, []); // empty dependency array will run effect only once (similar to componentDidMount)
 
   return (
     <Wrapper>
@@ -33,7 +22,6 @@ const App = () => {
       <QandA productId={productId} />
       <Divider />
       <Reviews productId={productId} />
-      <Footer />
     </Wrapper>
   );
 };
@@ -48,23 +36,13 @@ const Header = styled.div`
   padding: 5px;
 `;
 
-const Footer = styled.div`
-  background:orange;
-  height: 100px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  color: white;
-  padding: 5px;
-`;
-
 const Divider = styled.hr`
-display: block;
-height: 0px;
-border: 0;
-border-top: 2px solid #ccc;
-margin: 0;
-padding: 0;
+  display: block;
+  height: 1px;
+  border: 0;
+  border-top: 2px solid #ccc;
+  margin: 0;
+  padding: 0;
 `;
 
 const Wrapper = styled.div`
